Made a cool asteroid field
I made a scene that is a field of asteroid chunks that you can collide with and thus need to dodge to go fast. I added the breakapart asteroids prefractured instead of having the instantiate switch. I made prefabs for the particle field, clusters of prefractured asteroids, and an asteroid chunk exploder prefab that will explode out chunks of asteroids on start.